Move libSQL permit regression into client suite
